requirements.txt
flask>=3.0.0
requests>=2.31.0
aiohttp>=3.9.0
aiofiles>=23.0.0
python-dotenv>=1.0.0
rich>=13.0.0
edge-tts>=6.1.0
moviepy>=2.0.0
Pillow>=10.0.0
numpy>=2.0.0
yt-dlp>=2024.0.0
faster-whisper>=1.0.0
